+++ styles/chefOrders.scss
@@ -7,12 +7,12 @@ table, th, td {
 th, td {
   padding: 15px 100px;
 }
-.orders-title {
+.chef-orders-title {
   display:flex;
   justify-content: center;
 }
 
-.orders-table{
+.chef-orders-table{
   display:flex;
   justify-content: center;
 }
